change syntax for Python3

+++ alphalogic_api3/examples/example_simple.py
@@ -39,7 +39,7 @@ class MyRoot(Root):
             ]
 
     # Commands
-    @command(result_type=long)
+    @command(result_type=int)
     def cmd_alarm(self, where='here', when=datetime.datetime.now(), why=2):
         self.simple_event.emit()
         return self.param_count.val
